Add cypress tests
